fix: split views
