+++ src/containers/Tenant/Diagnostics/TopQueries/TopQueries.tsx
@@ -1,16 +1,12 @@
 import React from 'react';
 
-import type {Column} from '@gravity-ui/react-data-table';
-import type {RadioButtonOption, SelectOption} from '@gravity-ui/uikit';
-import {RadioButton, Select, TableColumnSetup} from '@gravity-ui/uikit';
+import type {RadioButtonOption} from '@gravity-ui/uikit';
+import {RadioButton} from '@gravity-ui/uikit';
 import {useHistory, useLocation} from 'react-router-dom';
 import {StringParam, useQueryParam} from 'use-query-params';
 import {z} from 'zod';
 
 import type {DateRangeValues} from '../../../../components/DateRange';
-import {DateRange} from '../../../../components/DateRange';
-import {Search} from '../../../../components/Search';
-import {TableWithControlsLayout} from '../../../../components/TableWithControlsLayout/TableWithControlsLayout';
 import {parseQuery} from '../../../../routes';
 import {setTopQueriesFilters} from '../../../../store/reducers/executeTopQueries/executeTopQueries';
 import type {TimeFrame} from '../../../../store/reducers/executeTopQueries/types';
@@ -20,29 +16,16 @@ import {
     TENANT_PAGES_IDS,
     TENANT_QUERY_TABS_ID,
 } from '../../../../store/reducers/tenant/constants';
-import {cn} from '../../../../utils/cn';
-import {useTypedDispatch, useTypedSelector} from '../../../../utils/hooks';
-import {useSelectedColumns} from '../../../../utils/hooks/useSelectedColumns';
+import {useTypedDispatch} from '../../../../utils/hooks';
 import {useChangeInputWithConfirmation} from '../../../../utils/hooks/withConfirmation/useChangeInputWithConfirmation';
 import {TenantTabsGroups, getTenantPath} from '../../TenantPages';
 
-import {RunningQueriesData} from './RunningQueriesData';
-import {TopQueriesData} from './TopQueriesData';
-import {getRunningQueriesColumns, getTopQueriesColumns} from './columns/columns';
-import {
-    DEFAULT_RUNNING_QUERIES_COLUMNS,
-    DEFAULT_TOP_QUERIES_COLUMNS,
-    REQUIRED_RUNNING_QUERIES_COLUMNS,
-    REQUIRED_TOP_QUERIES_COLUMNS,
-    RUNNING_QUERIES_SELECTED_COLUMNS_LS_KEY,
-    TOP_QUERIES_COLUMNS_TITLES,
-    TOP_QUERIES_SELECTED_COLUMNS_LS_KEY,
-} from './columns/constants';
+import {RunningQueriesContent} from './RunningQueriesContent';
+import {TopQueriesContent} from './TopQueriesContent';
+import {TimeFrameIds} from './constants';
 import i18n from './i18n';
 
 import './TopQueries.scss';
-
-const b = cn('kv-top-queries');
 
 const QueryModeIds = {
     top: 'top',
@@ -64,35 +47,8 @@ const QUERY_MODE_OPTIONS: RadioButtonOption[] = [
     },
 ];
 
-const TimeFrameIds = {
-    hour: 'hour',
-    minute: 'minute',
-} as const;
-
 const queryModeSchema = z.nativeEnum(QueryModeIds).catch(QueryModeIds.top);
 const timeFrameSchema = z.nativeEnum(TimeFrameIds).catch(TimeFrameIds.hour);
-
-const TIME_FRAME_OPTIONS: SelectOption[] = [
-    {
-        value: TimeFrameIds.hour,
-        content: i18n('timeframe_hour'),
-    },
-    {
-        value: TimeFrameIds.minute,
-        content: i18n('timeframe_minute'),
-    },
-];
-
-const DEFAULT_TIME_FILTER_VALUE = {
-    start: {
-        value: 'now-6h',
-        type: 'relative',
-    },
-    end: {
-        value: 'now',
-        type: 'relative',
-    },
-} as const;
 
 interface TopQueriesProps {
     tenantName: string;
@@ -109,8 +65,6 @@ export const TopQueries = ({tenantName}: TopQueriesProps) => {
     const timeFrame = timeFrameSchema.parse(_timeFrame);
 
     const isTopQueries = queryMode === QueryModeIds.top;
-
-    const filters = useTypedSelector((state) => state.executeTopQueries);
 
     const applyRowClick = React.useCallback(
         (input: string) => {
@@ -144,68 +98,28 @@ export const TopQueries = ({tenantName}: TopQueriesProps) => {
         dispatch(setTopQueriesFilters(value));
     };
 
-    // Get columns based on query mode
-    const columns: Column<any>[] = React.useMemo(() => {
-        return isTopQueries ? getTopQueriesColumns() : getRunningQueriesColumns();
-    }, [isTopQueries]);
+    const renderQueryModeControl = React.useCallback(() => {
+        return (
+            <RadioButton options={QUERY_MODE_OPTIONS} value={queryMode} onUpdate={setQueryMode} />
+        );
+    }, [queryMode, setQueryMode]);
 
-    // Use selected columns hook
-    const {columnsToShow, columnsToSelect, setColumns} = useSelectedColumns(
-        columns,
-        isTopQueries
-            ? TOP_QUERIES_SELECTED_COLUMNS_LS_KEY
-            : RUNNING_QUERIES_SELECTED_COLUMNS_LS_KEY,
-        TOP_QUERIES_COLUMNS_TITLES,
-        isTopQueries ? DEFAULT_TOP_QUERIES_COLUMNS : DEFAULT_RUNNING_QUERIES_COLUMNS,
-        isTopQueries ? REQUIRED_TOP_QUERIES_COLUMNS : REQUIRED_RUNNING_QUERIES_COLUMNS,
-    );
-
-    const DataComponent = isTopQueries ? TopQueriesData : RunningQueriesData;
-
-    return (
-        <TableWithControlsLayout>
-            <TableWithControlsLayout.Controls>
-                <RadioButton
-                    options={QUERY_MODE_OPTIONS}
-                    value={queryMode}
-                    onUpdate={setQueryMode}
-                />
-                {isTopQueries && (
-                    <Select
-                        options={TIME_FRAME_OPTIONS}
-                        value={[timeFrame]}
-                        onUpdate={handleTimeFrameChange}
-                    />
-                )}
-                {isTopQueries && (
-                    <DateRange
-                        from={filters.from}
-                        to={filters.to}
-                        onChange={handleDateRangeChange}
-                        defaultValue={DEFAULT_TIME_FILTER_VALUE}
-                    />
-                )}
-                <Search
-                    value={filters.text}
-                    onChange={handleTextSearchUpdate}
-                    placeholder={i18n('filter.text.placeholder')}
-                    className={b('search')}
-                />
-                <TableColumnSetup
-                    popupWidth={200}
-                    items={columnsToSelect}
-                    showStatus
-                    onUpdate={setColumns}
-                    sortable={false}
-                />
-            </TableWithControlsLayout.Controls>
-            <DataComponent
-                database={tenantName}
-                onRowClick={onRowClick}
-                rowClassName={b('row')}
-                timeFrame={timeFrame}
-                columns={columnsToShow}
-            />
-        </TableWithControlsLayout>
+    return isTopQueries ? (
+        <TopQueriesContent
+            tenantName={tenantName}
+            timeFrame={timeFrame}
+            renderQueryModeControl={renderQueryModeControl}
+            onRowClick={onRowClick}
+            handleTimeFrameChange={handleTimeFrameChange}
+            handleDateRangeChange={handleDateRangeChange}
+            handleTextSearchUpdate={handleTextSearchUpdate}
+        />
+    ) : (
+        <RunningQueriesContent
+            tenantName={tenantName}
+            renderQueryModeControl={renderQueryModeControl}
+            onRowClick={onRowClick}
+            handleTextSearchUpdate={handleTextSearchUpdate}
+        />
     );
 };
